src/options/requirements.txt
# src/options/requirements.txt
functions-framework
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-firestore
google-cloud-pubsub
google-cloud-storage
google-genai
matplotlib
mplfinance
pandas
pandas-gbq
pandas-ta
pyarrow
requests
tenacity
db-dtypes
sec-api
beautifulsoup4
